// examples-dj/workflow-example/packages/workflow-server/src/language-server/generated/ast.ts
export type Node = ActivityNode | Category | TaskNode;

export const Node = 'Node';
// ...
export type Weight = 'high' | 'low' | 'medium';

export function isWeight(item: unknown): item is Weight {
    return item === 'low' || item === 'medium' || item === 'high';
}

export interface ActivityNode extends AstNode {
    readonly $container: Model;
    readonly $type: 'ActivityNode';
    __id: string
    name: string
    nodeType?: NodeType
}

export const ActivityNode = 'ActivityNode';

export function isActivityNode(item: unknown): item is ActivityNode {
    return reflection.isInstance(item, ActivityNode);
}

export interface Category extends AstNode {
    readonly $container: Model;
    readonly $type: 'Category';
    __id: string
    children?: Model
    label?: string
    name: string
}

export const Category = 'Category';

export function isCategory(item: unknown): item is Category {
    return reflection.isInstance(item, Category);
}

export interface Edge extends AstNode {
    readonly $type: 'Edge' | 'WeightedEdge';
    __id: string
    source: Reference<Node>
    target: Reference<Node>
}

export const Edge = 'Edge';
// ...
export interface TaskNode extends AstNode {
    readonly $container: Model;
    readonly $type: 'TaskNode';
    __id: string
    duration?: number
    label?: string
    name: string
    reference?: string
    taskType?: TaskType
}

export const TaskNode = 'TaskNode';

export function isTaskNode(item: unknown): item is TaskNode {
    return reflection.isInstance(item, TaskNode);
}

export interface WeightedEdge extends Edge {
    readonly $type: 'WeightedEdge';
    __id: string
    source: Reference<Node>
    target: Reference<Node>
    weight: Weight
}

export const WeightedEdge = 'WeightedEdge';